requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
opencv-python==4.7.0.72
numpy==1.24.3
pydantic==1.10.7
python-magic==0.4.27
Pillow==9.5.0
scikit-image==0.20.0
filetype